hotfix 1

+++ triangle.rb
@@ -1,8 +1,8 @@
 # Программа для вычисления площади треугольника
 print "Длинна основания = "
-a_tria = gets.chomp.to_i  
+a_tria = gets.chomp.to_f  
 
 print "Высота треугольника = "
-h_tria = gets.chomp.to_i
+h_tria = gets.chomp.to_f
 
 puts "Площадь треугольника составляет #{(a_tria * h_tria * 0.5)}"
